fixed file names and conflicts from merging and finished level 2
