+++ Descension/Assets/Scripts/Environment/DescendHole.cs
@@ -14,7 +14,7 @@ namespace Environment
         void OnCollisionEnter2D(Collision2D collision) {
             if (collision.collider.tag == "Player") {
                 if (FindObjectOfType<PlayerController>().ropeQuantity > 0) {
-                    FindObjectOfType<PlayerController>().addRope(-1);
+                    FindObjectOfType<PlayerController>().AddRope(-1);
                     showText("Descend to level two...");
                     SceneManager.LoadScene(nextLevel);
                 } else {
