+++ Descension/Assets/Scripts/Environment/Message.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 690ce12001eaec347a9a979aaf11b73f
+guid: 41a26fbdb710428488f9ae31121917ab
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
