+++ Descension/Assets/Scripts/Items/arrowsItem.cs
@@ -11,7 +11,7 @@ namespace Items
         public float quantity = 20;
 
         void OnCollisionEnter2D(Collision2D collision) {
-            if (collision.gameObject.name == "Player") {
+            if (collision.collider.tag == "Player") {
                 FindObjectOfType<PlayerController>().AddArrows(this.quantity);
                 showText("Arrows collected");
                 Destroy(gameObject);
