+++ Descension/Assets/Scripts/Items/ropeItem.cs
@@ -11,7 +11,7 @@ namespace Items
         public float quantity = 1;
         
         void OnCollisionEnter2D(Collision2D collision) {
-            if (collision.gameObject.name == "Player") {
+            if (collision.collider.tag == "Player") {
                 FindObjectOfType<PlayerController>().AddRope(this.quantity);
                 showText("Rope Collected");
                 Destroy(gameObject);
